+++ lib/smart_core/types/value/rational.rb
@@ -1,13 +1,18 @@
 # frozen_string_literal: true
 
+using SmartCore::Ext::BasicObjectAsObject
+
 # @api public
-# @since 0.x.0
+# @since 0.9.0
 SmartCore::Types::Value.define_type(:Rational) do |type|
   type.define_checker do |value|
-    # TODO: realize
+    value.is_a?(::Rational)
   end
 
   type.define_caster do |value|
-    # TODO: realize
+    ::Kernel.Rational(value)
+  rescue ::TypeError, ::FloatDomainError
+    # NOTE: FloatDomainError is raised when you try to type cast Float::INFINITY or Float::NAN
+    raise(SmartCore::Types::TypeCastingError, 'Non-castable to Rational')
   end
 end